More the sentence-expression code to it's own file.
