+++ link-grammar/Makefile.am
@@ -142,6 +142,7 @@ liblink_grammar_la_SOURCES =        \
 	string-set.c                     \
 	string-id.c                      \
 	tokenize/anysplit.c              \
+	tokenize/lookup-exprs.c          \
 	tokenize/spellcheck-aspell.c     \
 	tokenize/spellcheck-hun.c        \
 	tokenize/tokenize.c              \
